# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "绦管后台 API 文档",
    "description": "电商后台管理系统RESTful API，支持小程序对接",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://qimengzhiyue.cn/api/v1",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "开发环境"
    }
  ],
...
  "paths": {
...
    "/orders": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "刕efa订单",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "items"
                ],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "product_id": {
                          "type": "integer"
                        },
                        "quantity": {
                          "type": "integer"
                        },
                        "price": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "product_id",
                        "quantity",
                        "price"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "刕efa成功"
          },
          "400": {
            "description": "参数无效"
          },
          "404": {
            "description": "商品不存在"
          }
        }
      }
    },
    "/orders/my": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "我的订单",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/orders/{orderId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "订单详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "不存在"
          }
        }
      }
    },
    "/orders/{orderId}/status": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "更新状态",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "400": {
            "description": "非法转换"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token 认证"
      }
    },
...
  }
}